create modal to add employee
