+++ src/App.js
@@ -1,27 +1,36 @@
-import logo from './logo.svg';
 import './App.css';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { ConfigProvider } from 'antd';
 
 import EmployeeContextProvider from './contexts/employeeContext';
 import TechnicalContextProvider from './contexts/technicalContext';
-import AddEmployee from './pages/employee/addEmployee';
+import Employees from './pages/employee/employees';
 
 function App() {
   return (
-    <EmployeeContextProvider>
-      <TechnicalContextProvider>
-        <div>
-          Sidebar
-        </div>
-        <div>
-          <BrowserRouter>
-            <Routes>
-              <Route path="/employee/create" element={<AddEmployee />} />
-            </Routes>
-          </BrowserRouter>
-        </div>
-      </TechnicalContextProvider>
-    </EmployeeContextProvider>
+    <ConfigProvider
+      theme={{
+        token: {
+          colorPrimary: '#3F51B5',
+          borderRadius: 2,
+        },
+      }}
+    >
+      <EmployeeContextProvider>
+        <TechnicalContextProvider>
+          <div>
+            Sidebar
+          </div>
+          <div>
+            <BrowserRouter>
+              <Routes>
+                <Route path="/employee" element={<Employees />} />
+              </Routes>
+            </BrowserRouter>
+          </div>
+        </TechnicalContextProvider>
+      </EmployeeContextProvider>
+    </ConfigProvider>
   );
 }
