+++ src/contexts/constants.js
@@ -1,3 +1,3 @@
-export const apiUrl = process.env.NODE_ENV !== 'production' ? 'http://localhost:4000' : 'url'
+// export const apiUrl = process.env.NODE_ENV !== 'production' ? 'http://localhost:4000' : 'url'
 
-// export const apiUrl = process.env.NODE_ENV !== 'production' ? 'https://ojt-7t4-server.onrender.com' : 'url'
+export const apiUrl = process.env.NODE_ENV !== 'production' ? 'https://ojt-7t4-server.onrender.com' : 'url'
